Consistent `Example 2` in Router State section
If this moves the documentation away from the direction you want it - no worries. That said, as I was scanning that table, I was confused by the fact that Example 1 clearly stayed constant through its whole column, but Example 2 seemed to change.

If the old values for Example 2 are in fact better, I think its probably worth explaining why.